+++ src/app/governance/page.tsx
@@ -14,7 +14,7 @@ import { PageLayout } from '@/components/PageLayout'
 import { PageSection } from '@/components/PageSection'
 import { StructuredDataScript } from '@/components/StructuredDataScript'
 
-import { CalendarCards } from './components/CalendarCards'
+import { CalendarCards } from './components/CalendarCards/CalendarCards'
 import { CTAPageSection } from './components/CTAPageSection'
 import { governanceDocsData } from './data/governanceDocsData'
 import { generateStructuredData } from './utils/generateStructuredData'
